manageapp/requirements.txt
asgiref==3.8.1
astroid==3.3.8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
Django==5.1.5
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
isort==5.13.2
jwcrypto==1.5.6
mccabe==0.7.0
mysqlclient==2.2.7
oauth2_provider==0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.3
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
views==0.3
